More work on the planner, moving around packages
