+++ pkg/catalogue/paths.go
@@ -1,12 +1,10 @@
-package runner
+package catalogue
 
 import (
 	"fmt"
 	"os"
 	"path"
 	"strings"
-
-	"github.com/projectdiscovery/gologger"
 )
 
 // isRelative checks if a given path is a relative path
@@ -22,7 +20,7 @@ func isRelative(filePath string) bool {
 // looking in the current directory or checking the nuclei templates directory.
 //
 // Current directory is given preference over the nuclei-templates directory.
-func (r *Runner) resolvePath(templateName string) (string, error) {
+func (c *Catalogue) resolvePath(templateName string) (string, error) {
 	curDirectory, err := os.Getwd()
 	if err != nil {
 		return "", err
@@ -30,19 +28,22 @@ func (r *Runner) resolvePath(templateName string) (string, error) {
 
 	templatePath := path.Join(curDirectory, templateName)
 	if _, err := os.Stat(templatePath); !os.IsNotExist(err) {
-		gologger.Debugf("Found template in current directory: %s\n", templatePath)
-
 		return templatePath, nil
 	}
 
-	if r.templatesConfig != nil {
-		templatePath := path.Join(r.templatesConfig.TemplatesDirectory, templateName)
+	if c.templatesDirectory != "" {
+		templatePath := path.Join(c.templatesDirectory, templateName)
 		if _, err := os.Stat(templatePath); !os.IsNotExist(err) {
-			gologger.Debugf("Found template in nuclei-templates directory: %s\n", templatePath)
-
 			return templatePath, nil
 		}
 	}
+	return "", fmt.Errorf("no such path found: %s", templateName)
+}
 
+func resolvePathWithBaseFolder(baseFolder, templateName string) (string, error) {
+	templatePath := path.Join(baseFolder, templateName)
+	if _, err := os.Stat(templatePath); !os.IsNotExist(err) {
+		return templatePath, nil
+	}
 	return "", fmt.Errorf("no such path found: %s", templateName)
 }
