+++ pkg/planner/planner.go
@@ -1,20 +1,32 @@
 package planner
 
 import (
-	"github.com/projectdiscovery/nuclei/v2/pkg/quarks"
-	"github.com/projectdiscovery/nuclei/v2/pkg/quarks/input"
+	"github.com/projectdiscovery/nuclei/v2/pkg/catalogue"
+	"github.com/projectdiscovery/nuclei/v2/pkg/quarks/requests/http"
 )
 
 // ExecutionPlan is the execution plan of a certain scan input.
 type ExecutionPlan struct {
 	// Flow is the scan flow defined in the form a quark execution graph.
-	Flow []quarks.CompiledQuark
+	//Inputs
+	httpRequests []*http.CompiledRequest
 }
 
-// PlanExecution plans the execution flow of a scan using input quarks.
-func PlanExecution(quarks []input.Input) (*ExecutionPlan, error) {
-	for _, quark := range quarks {
+// ExecutionStep defines a single step to be performed during execution
+type ExecutionStep struct {
+}
 
+// Plan plans the execution flow of a scan using input quarks.
+func Plan(inputs []*catalogue.CompiledInput) (*ExecutionPlan, error) {
+	plan := &ExecutionPlan{}
+
+	for _, input := range inputs {
+		if input.Type == catalogue.TemplateInputType {
+			plan.planTemplate(input)
+		}
 	}
-	// Compile all the quarks into
+}
+
+// planTemplate plans a template into the execution plan
+func (e *ExecutionPlan) planTemplate(input *catalogue.CompiledInput) {
 }
